+++ packages/anipres/src/Anipres.tsx
@@ -9,15 +9,15 @@ import {
   DefaultKeyboardShortcutsDialog,
   DefaultKeyboardShortcutsDialogContent,
   uniqueId,
-  react,
   useAtom,
   useValue,
+  react,
 } from "tldraw";
 import type {
+  Atom,
   TLUiOverrides,
   TLComponents,
   Editor,
-  TLShape,
   TldrawProps,
   TLStoreSnapshot,
   TLEditorSnapshot,
@@ -26,22 +26,20 @@ import type {
 } from "tldraw";
 import "tldraw/tldraw.css";
 
-import { SlideShapeType, SlideShapeUtil } from "./SlideShapeUtil";
+import { SlideShapeType } from "./SlideShapeUtil";
 import { SlideShapeTool } from "./SlideShapeTool";
 import { ControlPanel } from "./ControlPanel";
 import { createModeAwareDefaultComponents } from "./mode-aware-components";
 import {
-  runStep,
   cueFrameToJsonObject,
   getFrame,
   type CameraZoomFrameAction,
   type CueFrame,
   type SubFrame,
-  reconcileShapeDeletion,
   getSubFrame,
   subFrameToJsonObject,
 } from "./models";
-import { EditorSignals } from "./editor-signals";
+import { PresentationManager } from "./presentation-manager";
 import React, {
   useCallback,
   useEffect,
@@ -52,7 +50,7 @@ import React, {
 
 import "./tldraw-overrides.css";
 
-const customShapeUtils = [SlideShapeUtil];
+import { customShapeUtils } from "./shape-utils";
 const customTools = [SlideShapeTool];
 
 // We use atoms as it's Tldraw's design,
@@ -66,30 +64,25 @@ function usePerInstanceAtoms() {
     true,
   );
   const $presentationMode = useAtom<boolean>("presentation mode", false);
-  const $currentStepIndex = useAtom<number>("current step index", 0);
 
   return useMemo(() => {
     return {
       $stepHotkeyEnabled,
       $presentationModeHotkeyEnabled,
       $presentationMode,
-      $currentStepIndex,
     };
-  }, [
-    $stepHotkeyEnabled,
-    $presentationModeHotkeyEnabled,
-    $presentationMode,
-    $currentStepIndex,
-  ]);
+  }, [$stepHotkeyEnabled, $presentationModeHotkeyEnabled, $presentationMode]);
 }
 export type AnipresAtoms = ReturnType<typeof usePerInstanceAtoms>;
 
-const makeUiOverrides = ({
-  $stepHotkeyEnabled,
-  $presentationModeHotkeyEnabled,
-  $currentStepIndex,
-  $presentationMode,
-}: AnipresAtoms): TLUiOverrides => {
+const makeUiOverrides = (
+  {
+    $stepHotkeyEnabled,
+    $presentationModeHotkeyEnabled,
+    $presentationMode,
+  }: AnipresAtoms,
+  presentationManagerRef: React.RefObject<PresentationManager>,
+): TLUiOverrides => {
   return {
     actions(_, actions) {
       actions["next-step"] = {
@@ -101,7 +94,12 @@ const makeUiOverrides = ({
             return;
           }
 
-          $currentStepIndex.update((v) => v + 1);
+          const presentationManager = presentationManagerRef.current;
+          if (presentationManager == null) {
+            return;
+          }
+
+          presentationManager.moveTo((v) => v + 1);
         },
       };
 
@@ -114,7 +112,12 @@ const makeUiOverrides = ({
             return;
           }
 
-          $currentStepIndex.update((v) => v - 1);
+          const presentationManager = presentationManagerRef.current;
+          if (presentationManager == null) {
+            return;
+          }
+
+          presentationManager.moveTo((v) => v - 1);
         },
       };
 
@@ -163,16 +166,20 @@ const makeUiOverrides = ({
 };
 
 const createComponents = (
-  $editorSignalsRef: React.RefObject<EditorSignals | null>,
-  { $currentStepIndex, $presentationMode }: AnipresAtoms,
+  presentationManagerRef: React.RefObject<PresentationManager | null>,
+  signals: {
+    $currentStepIndex: Atom<number>;
+    $presentationMode: Atom<boolean>;
+  },
 ): TLComponents => {
+  const { $currentStepIndex, $presentationMode } = signals;
   return {
     TopPanel: () => {
       const editor = useEditor();
-      const $editorSignals = $editorSignalsRef.current;
+      const presentationManager = presentationManagerRef.current;
       const presentationMode = useValue($presentationMode);
       const currentStepIndex = useValue($currentStepIndex);
-      if ($editorSignals == null) {
+      if (presentationManager == null) {
         return null;
       }
       if (presentationMode) {
@@ -181,10 +188,10 @@ const createComponents = (
       return (
         <ControlPanel
           editor={editor}
-          $editorSignals={$editorSignals}
+          presentationManager={presentationManager}
           currentStepIndex={currentStepIndex}
           onCurrentStepIndexChange={(newIndex) => {
-            $currentStepIndex.set(newIndex);
+            presentationManager.moveTo(newIndex);
           }}
           onPresentationModeEnter={() => {
             $presentationMode.set(true);
@@ -223,7 +230,7 @@ const createComponents = (
 interface InnerProps {
   onMount: (
     editor: Editor,
-    $editorSignals: EditorSignals,
+    presentationManager: PresentationManager,
   ) => (() => void) | void;
   snapshot?: TLEditorSnapshot | TLStoreSnapshot;
   perInstanceAtoms: AnipresAtoms;
@@ -232,11 +239,16 @@ interface InnerProps {
 const Inner = (props: InnerProps) => {
   const { onMount, snapshot, perInstanceAtoms, assetUrls } = props;
 
-  const $editorSignalsRef = useRef<EditorSignals | null>(null);
+  const presentationManagerRef = useRef<PresentationManager | null>(null);
+
+  const $currentStepIndex = useAtom<number>("current step index", 0);
 
   const handleMount = (editor: Editor) => {
-    const $editorSignals = new EditorSignals(editor);
-    $editorSignalsRef.current = $editorSignals;
+    const presentationManager = PresentationManager.create(
+      editor,
+      $currentStepIndex,
+    );
+    presentationManagerRef.current = presentationManager;
 
     const stopHandlers: (() => void)[] = [];
 
@@ -244,7 +256,7 @@ const Inner = (props: InnerProps) => {
       editor.sideEffects.registerBeforeCreateHandler("shape", (shape) => {
         if (shape.type === SlideShapeType && shape.meta?.frame == null) {
           // Auto attach camera cueFrame to the newly created slide shape
-          const orderedSteps = $editorSignals.getOrderedSteps();
+          const orderedSteps = presentationManager.$getOrderedSteps();
           const lastCameraCueFrame = orderedSteps
             .reverse()
             .flat()
@@ -309,7 +321,7 @@ const Inner = (props: InnerProps) => {
     );
     stopHandlers.push(
       editor.sideEffects.registerAfterDeleteHandler("shape", (shape) => {
-        reconcileShapeDeletion(editor, $editorSignals, shape);
+        presentationManager.reconcileShapeDeletion(shape);
       }),
     );
 
@@ -385,109 +397,43 @@ const Inner = (props: InnerProps) => {
       }
     });
 
-    react("current step index", () => {
-      const index = perInstanceAtoms.$currentStepIndex.get();
+    stopHandlers.push(
+      react("turn off edit tools in presentation mode", () => {
+        const presentationMode = perInstanceAtoms.$presentationMode.get();
+        if (presentationMode) {
+          editor.setCurrentTool("select");
+        }
+      }),
+    );
 
-      setTimeout(() => {
-        // `runStep` internally refers to another reactive value that depends on the `editor` object.
-        // So we need to put `runStep()` here to avoid an infinite loop.
-        const orderedSteps = $editorSignals.getOrderedSteps();
-        runStep(editor, orderedSteps, index);
-      });
-    });
-
-    onMount?.(editor, $editorSignals);
+    onMount?.(editor, presentationManager);
 
     return () => {
       stopHandlers.forEach((stopHandler) => stopHandler());
     };
   };
 
-  const determineShapeHidden = (shape: TLShape): boolean => {
+  const determineShapeVisibility: TldrawProps["getShapeVisibility"] = (
+    shape,
+    editor,
+  ) => {
     const presentationMode = perInstanceAtoms.$presentationMode.get();
-    const editMode = !presentationMode;
-    const HIDDEN = true;
-    const SHOW = false;
-    if (editMode) {
-      return SHOW;
+    if (!presentationMode) {
+      return "visible";
     }
 
-    if (shape.type === SlideShapeType) {
-      return HIDDEN;
-    }
+    // This callback can be called before `onMount` is called and the refs are set.
+    // So we need to get presentationManager here using the editor object passed to this callback
+    // instead of relying on the refs that are set in `onMount`.
+    // `presentationManager.create` ensures that the same instance is returned for the same editor.
+    const presentationManager = PresentationManager.create(
+      editor,
+      $currentStepIndex,
+    );
 
-    if (shape.meta?.hiddenDuringAnimation) {
-      return HIDDEN;
-    }
-
-    const frame = getFrame(shape);
-    if (frame == null) {
-      // No animation frame is attached to this shape, so it should always be visible
-      return SHOW;
-    }
-
-    if ($editorSignalsRef.current == null) {
-      // Fallback: If editorSignalsRef.current is null, assume the shape should be hidden
-      return HIDDEN;
-    }
-
-    const orderedSteps = $editorSignalsRef.current.getOrderedSteps();
-    const currentStepIndex = perInstanceAtoms.$currentStepIndex.get();
-
-    // The last frame of a finished animation should always be visible
-    if (frame.type === "cue") {
-      const cueFrame = frame;
-      const isFuture = cueFrame.globalIndex > currentStepIndex;
-      if (isFuture) {
-        return HIDDEN;
-      }
-
-      const lastBatchIncludingThisTrack = orderedSteps
-        .slice(0, currentStepIndex + 1)
-        .reverse()
-        .flat()
-        .find((ab) => ab.trackId === cueFrame.trackId);
-      const isLatestPrevInTrack =
-        lastBatchIncludingThisTrack &&
-        lastBatchIncludingThisTrack.data.findIndex(
-          (frame) => frame.id === cueFrame.id,
-        ) ===
-          lastBatchIncludingThisTrack.data.length - 1;
-      if (isLatestPrevInTrack) {
-        return SHOW;
-      }
-    } else if (frame.type === "sub") {
-      const subFrame = frame;
-      const thisBatch = orderedSteps
-        .flat()
-        .find((ab) => ab.data.some((frame) => frame.id === subFrame.id));
-      if (thisBatch == null) {
-        // This should never happen, but just in case
-        return HIDDEN;
-      }
-
-      const isFuture = thisBatch.globalIndex > currentStepIndex;
-      if (isFuture) {
-        return HIDDEN;
-      }
-
-      const lastBatchIncludingThisTrack = orderedSteps
-        .slice(0, currentStepIndex + 1)
-        .reverse()
-        .flat()
-        .find((ab) => ab.trackId === thisBatch.trackId);
-      const isLatestPrevInTrack =
-        lastBatchIncludingThisTrack &&
-        lastBatchIncludingThisTrack.data.findIndex(
-          (frame) => frame.id === subFrame.id,
-        ) ===
-          lastBatchIncludingThisTrack.data.length - 1;
-      if (isLatestPrevInTrack) {
-        return SHOW;
-      }
-    }
-
-    return HIDDEN;
+    const shapeVisibilities =
+      presentationManager.$getShapeVisibilitiesInPresentationMode();
+    return shapeVisibilities[shape.id] ?? "hidden";
   };
 
   return (
@@ -495,12 +441,15 @@ const Inner = (props: InnerProps) => {
       onMount={handleMount}
       components={{
         ...createModeAwareDefaultComponents(perInstanceAtoms.$presentationMode),
-        ...createComponents($editorSignalsRef, perInstanceAtoms),
+        ...createComponents(presentationManagerRef, {
+          $currentStepIndex,
+          $presentationMode: perInstanceAtoms.$presentationMode,
+        }),
       }}
-      overrides={makeUiOverrides(perInstanceAtoms)}
+      overrides={makeUiOverrides(perInstanceAtoms, presentationManagerRef)}
       shapeUtils={customShapeUtils}
       tools={customTools}
-      isShapeHidden={determineShapeHidden}
+      getShapeVisibility={determineShapeVisibility}
       options={{
         maxPages: 1,
       }}
@@ -515,11 +464,7 @@ const MemoizedInner = React.memo(Inner);
 
 export interface AnipresProps {
   presentationMode?: boolean;
-  onMount?: (
-    editor: Editor,
-    $editorSignals: EditorSignals,
-    anipresAtoms: AnipresAtoms,
-  ) => void;
+  onMount?: (editor: Editor, moveTo: (stepIndex: number) => void) => void;
   snapshot?: InnerProps["snapshot"];
   assetUrls?: InnerProps["assetUrls"];
   stepHotkeyEnabled?: boolean;
@@ -539,7 +484,6 @@ export const Anipres = React.forwardRef<AnipresRef, AnipresProps>(
 
     const anipresAtoms = usePerInstanceAtoms();
     const {
-      $currentStepIndex,
       $presentationMode,
       $stepHotkeyEnabled,
       $presentationModeHotkeyEnabled,
@@ -553,38 +497,36 @@ export const Anipres = React.forwardRef<AnipresRef, AnipresProps>(
       $presentationModeHotkeyEnabled.set(presentationMode == null);
     }, [$presentationModeHotkeyEnabled, presentationMode]);
 
-    const editorAndSignalsRef = useRef<{
-      editor: Editor;
-      $editorSignals: EditorSignals;
-    } | null>(null);
-    const handleMount = useCallback(
-      (editor: Editor, $editorSignals: EditorSignals) => {
-        editorAndSignalsRef.current = {
-          editor,
-          $editorSignals: $editorSignals,
-        };
-        onMount?.(editor, $editorSignals, anipresAtoms);
-      },
-      [onMount, anipresAtoms],
-    );
-
     useEffect(() => {
       if (presentationMode != null) {
         $presentationMode.set(presentationMode);
       }
     }, [$presentationMode, presentationMode]);
 
+    const editorAndSignalsRef = useRef<{
+      editor: Editor;
+      presentationManager: PresentationManager;
+    } | null>(null);
+    const handleMount = useCallback(
+      (editor: Editor, presentationManager: PresentationManager) => {
+        editorAndSignalsRef.current = {
+          editor,
+          presentationManager,
+        };
+        onMount?.(editor, (stepIndex: number) => {
+          presentationManager.moveTo(stepIndex);
+        });
+      },
+      [onMount],
+    );
+
     useImperativeHandle(ref, () => ({
       rerunStep: () => {
         if (editorAndSignalsRef.current == null) {
           return;
         }
-        const { editor, $editorSignals } = editorAndSignalsRef.current;
-        runStep(
-          editor,
-          $editorSignals.getOrderedSteps(),
-          $currentStepIndex.get(),
-        );
+        const { presentationManager } = editorAndSignalsRef.current;
+        presentationManager.rerunStep();
       },
     }));