+++ packages/anipres/src/models.ts
@@ -1,10 +1,6 @@
-import { EASINGS, GroupShapeUtil, createShapeId, uniqueId } from "tldraw";
+import { EASINGS, GroupShapeUtil, uniqueId } from "tldraw";
 import type { Editor, JsonObject, TLShape, TLShapeId } from "tldraw";
-import {
-  OrderedTrackItem,
-  reassignGlobalIndexInplace,
-} from "./ordered-track-item";
-import { EditorSignals } from "./editor-signals";
+import { OrderedTrackItem } from "./ordered-track-item";
 
 export interface FrameActionBase extends JsonObject {
   type: string;
@@ -55,6 +51,8 @@ type BatchedFrames<T extends FrameAction = FrameAction> = [
 export type FrameBatch<T extends FrameAction = FrameAction> = OrderedTrackItem<
   BatchedFrames<T>
 >;
+
+export type Step = FrameBatch[];
 
 export function cueFrameToJsonObject(cf: CueFrame): JsonObject {
   const { id, type, globalIndex, trackId, action } = cf;
@@ -139,7 +137,7 @@ export function getNextGlobalIndexFromCueFrames(cueFrames: CueFrame[]): number {
   return globalIndexes.length > 0 ? Math.max(...globalIndexes) + 1 : 0;
 }
 
-export function getNextGlobalIndex(editor: Editor): number {
+function getNextGlobalIndex(editor: Editor): number {
   const shapes = editor.getCurrentPageShapes();
   const allCueFrames = shapes
     .map(getCueFrame)
@@ -198,6 +196,10 @@ export function getFrame(shape: TLShape): Frame | undefined {
   return getCueFrame(shape) ?? getSubFrame(shape);
 }
 
+export function getFrames(shapes: TLShape[]): Frame[] {
+  return shapes.map(getFrame).filter((frame) => frame != null);
+}
+
 export function getCueFrame(shape: TLShape): CueFrame | undefined {
   return isJsonObject(shape.meta.frame) && shape.meta.frame.type === "cue"
     ? jsonObjectToCueFrame(shape.meta.frame)
@@ -244,239 +246,4 @@ export function getFrameBatches(frames: Frame[]): FrameBatch[] {
   }
 
   return frameBatches;
-}
-
-export function getFramesFromFrameBatches(frameBatches: FrameBatch[]): Frame[] {
-  return frameBatches.flatMap((batch) => batch.data);
-}
-
-export type Step = FrameBatch[];
-
-export function getShapeByFrameId(
-  editor: Editor,
-  frameId: Frame["id"],
-): TLShape | undefined {
-  const shapes = editor.getCurrentPageShapes();
-  return shapes.find((shape) => getFrame(shape)?.id === frameId);
-}
-
-export function reconcileShapeDeletion(
-  editor: Editor,
-  $editorSignals: EditorSignals,
-  deletedShape: TLShape,
-) {
-  const deletedFrame = getFrame(deletedShape);
-  if (deletedFrame == null) {
-    return;
-  }
-
-  if (deletedFrame.type === "cue") {
-    // Reassign globalIndex
-    const steps = $editorSignals.getOrderedSteps();
-    reassignGlobalIndexInplace(steps);
-    steps.forEach((stepFrameBatches) => {
-      stepFrameBatches.forEach((frameBatch) => {
-        const newGlobalIndex = frameBatch.globalIndex;
-        const cueFrame = frameBatch.data[0];
-        const shape = getShapeByFrameId(editor, cueFrame.id);
-        if (shape == null) {
-          return;
-        }
-        editor.updateShape({
-          id: shape.id,
-          type: shape.type,
-          meta: {
-            frame: cueFrameToJsonObject({
-              ...cueFrame,
-              globalIndex: newGlobalIndex,
-            }),
-          },
-        });
-      });
-    });
-  } else if (deletedFrame.type === "sub") {
-    // Reassign prevFrameId
-    const shapes = editor.getCurrentPageShapes();
-    const allSubFrames = shapes
-      .map((shape) => ({ shape, subFrame: getSubFrame(shape) }))
-      .filter(({ subFrame }) => subFrame != null) as {
-      shape: TLShape;
-      subFrame: SubFrame;
-    }[];
-    allSubFrames.forEach(({ shape, subFrame }) => {
-      if (subFrame.prevFrameId === deletedFrame.id) {
-        editor.updateShape({
-          id: shape.id,
-          type: shape.type,
-          meta: {
-            frame: subFrameToJsonObject({
-              ...subFrame,
-              prevFrameId: deletedFrame.prevFrameId,
-            }),
-          },
-        });
-      }
-    });
-  }
-}
-
-async function runFrames(
-  editor: Editor,
-  frames: Frame[],
-  predecessorShape: TLShape | null,
-  historyStoppingPoint: string,
-): Promise<void> {
-  for (const frame of frames) {
-    const shape = getShapeByFrameId(editor, frame.id);
-    if (shape == null) {
-      throw new Error(`Shape not found for frame ${frame.id}`);
-    }
-
-    const action = frame.action;
-
-    const { duration = 0, easing = "easeInCubic" } = action;
-    const immediate = duration === 0;
-
-    if (action.type === "cameraZoom") {
-      const { inset = 0 } = action;
-
-      editor.stopCameraAnimation();
-      const bounds = editor.getShapePageBounds(shape);
-      if (!bounds) {
-        throw new Error(`Bounds not found for shape ${shape.id}`);
-      }
-      editor.selectNone();
-      editor.zoomToBounds(bounds, {
-        inset,
-        immediate,
-        animation: { duration, easing: EASINGS[easing] },
-      });
-    } else if (action.type === "shapeAnimation") {
-      editor.selectNone();
-
-      if (predecessorShape == null) {
-        predecessorShape = shape;
-        continue;
-      }
-
-      // Create and manipulate a temporary shape for animation
-      const animeShapeId = createShapeId();
-      editor.run(
-        () => {
-          editor.createShape({
-            ...predecessorShape,
-            id: animeShapeId,
-            type: shape.type,
-            meta: undefined,
-          });
-        },
-        { history: "ignore", ignoreShapeLock: true },
-      );
-
-      // HACK: Changes made by editor.animateShape() can't be ignored by `editor.run(..., { history: "ignore" })`
-      // because it's done in the `tick` event listener that is executed after the `editor.run()` returns.
-      // So we need to cancel the history records in another `tick` event listener manually.
-      const onTick = () => {
-        editor.bailToMark(historyStoppingPoint);
-      };
-      editor.on("tick", onTick);
-      editor.animateShape(
-        {
-          ...shape,
-          id: animeShapeId,
-          meta: undefined,
-        },
-        {
-          immediate,
-          animation: {
-            duration,
-            easing: EASINGS[easing],
-          },
-        },
-      );
-
-      setTimeout(() => {
-        editor.run(
-          () => {
-            editor.deleteShape(animeShapeId);
-          },
-          { history: "ignore", ignoreShapeLock: true },
-        );
-        editor.off("tick", onTick);
-      }, duration);
-    }
-
-    await new Promise((resolve) => setTimeout(resolve, duration));
-
-    predecessorShape = shape;
-  }
-}
-
-export function runStep(editor: Editor, steps: Step[], index: number): boolean {
-  const step = steps[index];
-  if (step == null) {
-    return false;
-  }
-
-  const markBeforeAnimation = editor.markHistoryStoppingPoint();
-
-  step.forEach((frameBatch) => {
-    const predecessorFrameBatch = steps
-      .slice(0, index)
-      .reverse()
-      .flat()
-      .find((fb) => fb.trackId === frameBatch.trackId);
-    const predecessorLastFrame = predecessorFrameBatch?.data.at(-1);
-    const predecessorShape =
-      predecessorLastFrame != null
-        ? getShapeByFrameId(editor, predecessorLastFrame.id)
-        : null;
-
-    const frames = frameBatch.data;
-    const frameShapes = frames
-      .map((frame) => getShapeByFrameId(editor, frame.id))
-      .filter((shape) => shape != null);
-
-    editor.run(
-      () => {
-        for (const shape of frameShapes) {
-          editor.updateShape({
-            id: shape.id,
-            type: shape.id,
-            meta: {
-              ...shape.meta,
-              hiddenDuringAnimation: true,
-            },
-          });
-        }
-      },
-      { history: "ignore", ignoreShapeLock: true },
-    );
-
-    runFrames(
-      editor,
-      frames,
-      predecessorShape ?? null,
-      markBeforeAnimation,
-    ).finally(() => {
-      editor.run(
-        () => {
-          for (const shape of frameShapes) {
-            editor.updateShape({
-              id: shape.id,
-              type: shape.id,
-              meta: {
-                ...shape.meta,
-                hiddenDuringAnimation: false,
-              },
-            });
-          }
-        },
-        { history: "ignore", ignoreShapeLock: true },
-      );
-      editor.bailToMark(markBeforeAnimation);
-    });
-  });
-
-  return true;
 }